Start using typescript
- Add rxjs and remove rxjs-es to reduce possible sources of confusion

It is working :D

Only the index file is ts, rest of the codebase is still js, but it
works. Only quirk is that webpack isn't watching the changes. It fails
on first change. I haven't done any further development with typescript
so far. Will wait for the watch issue to get resolved.

Let's see how this ride goes baby!