Remove redux-immutable
react-router-redux don't work well when using immutable state. The
specific issue was:
Using <IndexRedirect to='' /> would change the state, but it would not
reflect in the app. It would show a blank page and won't even attempt to
render App.js, and won't change the location. The location would change
for a microsecond, and then change back to `/`

- Redirect from `/` to `/date/<today>/`