Remove Dropbox support from this version
From the feedback I got about the app, it appear the whole react + redux
+ rx setup is somewhat overwhelming to someone new to redux (most
testers were aquainted with React). So I have decided to not include
Dropbox (hence its own set of complexities) and keep the app simple. May
be we can add offline support, but that is trivial, and I am leaving it
to be done at end of tutorial branch.

We're done here folks. Over to the tutorial branch!